Transferred plotting functions from the shap_result.py to the wandb_utils.py to better handle wandb importing necessity.
